+++ src/OrderLinesList/OrderLinesList.js
@@ -21,7 +21,7 @@ import {
   ResetButton,
   ResultsPane,
   SingleSearchForm,
-  useLocationFilters,
+  useLocalStorageFilters,
   useLocationSorting,
   useToggle,
 } from '@folio/stripes-acq-components';
@@ -70,7 +70,7 @@ function OrderLinesList({
     resetFilters,
     changeIndex,
     searchIndex,
-  ] = useLocationFilters(location, history, resetData);
+  ] = useLocalStorageFilters('OrderLinesList/filters', location, history, resetData);
   const [
     sortingField,
     sortingDirection,